Feat: Use open ai module for custom OpenAI url
