+++ package.json
@@ -12,12 +12,11 @@
     "lint:fix": "eslint . --ext .ts --fix",
     "format": "prettier --write \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
     "format:check": "prettier --check \"src/**/*.{js,jsx,ts,tsx,json,md}\"",
-    "postbuild": "chmod +x dist/index.js",
     "changelog": "conventional-changelog -p angular -i CHANGELOG.md -s -r 0",
     "version": "conventional-changelog -p angular -i CHANGELOG.md -s && git add CHANGELOG.md",
     "test": "npm run build && node dist/index.js -i README.md -o README-zh.md -l Chinese",
-    "test:ko": "npm run build && node dist/index.js -i README.md -o README-ko.md -l Korean --locale ko",
-    "test:ko-prompt": "npm run build && node dist/index.js -i src/prompts/system.md -o prompts_translated/system-ko.md -l Korean --locale ko"
+    "test:ko": "npm run build && node dist/index.js -i README.md -o README-ko.md -l Korean --locale ko --log",
+    "test:ko-prompt": "npm run build && node dist/index.js -i src/prompts/system.md -o prompts_translated/system-ko.md -l Korean --locale ko --log"
   },
   "bin": {
     "ai-markdown-translator": "dist/index.js"
@@ -35,6 +34,7 @@
     "axios": "^1.4.0",
     "dotenv": "^16.0.3",
     "glob": "^11.0.0",
+    "openai": "^4.89.0",
     "yargs": "^17.7.2"
   },
   "devDependencies": {
